requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# Oracle database connectivity
cx_Oracle==8.3.0
oracledb==1.4.2

# Additional data processing libraries
pandas==2.0.3
sqlalchemy==1.4.53

# Airflow Oracle provider
apache-airflow-providers-oracle==3.10.1
